+++ Changelog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Implement Crawl
 - Ignore Archive Url's
+- Splash to Fetch dynamic generated content 
 
 ## [0.0.1]
 ### Added